+++ api/sql/reset-test-db.sql
@@ -1,0 +1,2 @@
+DROP DATABASE IF EXISTS `uberlike_test`;
+CREATE DATABASE IF NOT EXISTS `uberlike_test` COLLATE utf8_unicode_ci;